- added option --targetstudio which is used to specify a program   which creates the target device node to create the image on.   In contrast to the targetdevice option there will be no image   file dumped from the specified device. This option is mainly   used in SuSE Studio

+++ modules/KIWIImage.pm
@@ -770,6 +770,9 @@ sub createImageEXT {
 	if (! defined $name) {
 		return;
 	}
+	if ($this->{targetDevice}) {
+		$device = $this->{targetDevice};
+	}
 	#==========================================
 	# Create filesystem on extend
 	#------------------------------------------
@@ -840,6 +843,9 @@ sub createImageReiserFS {
 	if (! defined $name) {
 		return;
 	}
+	if ($this->{targetDevice}) {
+		$device = $this->{targetDevice};
+	}
 	#==========================================
 	# Create filesystem on extend
 	#------------------------------------------
@@ -872,6 +878,9 @@ sub createImageBTRFS {
 	my $name = $this -> preImage ($device);
 	if (! defined $name) {
 		return;
+	}
+	if ($this->{targetDevice}) {
+		$device = $this->{targetDevice};
 	}
 	if (defined $rename) {
 		my $data = qxx (
@@ -916,6 +925,9 @@ sub createImageXFS {
 	my $name = $this -> preImage ($device);
 	if (! defined $name) {
 		return;
+	}
+	if ($this->{targetDevice}) {
+		$device = $this->{targetDevice};
 	}
 	#==========================================
 	# Create filesystem on extend
@@ -3368,14 +3380,33 @@ sub buildLogicalExtend {
 		#==========================================
 		# Create logical extend storage and FS
 		#------------------------------------------
-		unlink ($out);
-		my $data = qxx ("qemu-img create $out $seek 2>&1");
-		my $code = $? >> 8;
-		if ($code != 0) {
-			$kiwi -> error  ("Couldn't create logical extend");
-			$kiwi -> failed ();
-			$kiwi -> error  ($data);
-			return;
+		if (-x $this->{gdata}->{StudioNode}) {
+			#==========================================
+			# Call custom image creation tool...
+			#------------------------------------------
+			my $data = qxx ("$this->{gdata}->{StudioNode} $seek 2>&1");
+			my $code = $? >> 8;
+			chomp $data;
+			if (($code != 0) || (! -b $data)) {
+				$kiwi -> error  ("Failed creating Studio storage device: $data");
+				$kiwi -> failed ();
+				return;
+			}
+			$device = $data;
+			$this->{targetDevice} = $device;
+		} else {
+			#==========================================
+			# loop setup a disk device as file...
+			#------------------------------------------
+			unlink ($out);
+			my $data = qxx ("qemu-img create $out $seek 2>&1");
+			my $code = $? >> 8;
+			if ($code != 0) {
+				$kiwi -> error  ("Couldn't create logical extend");
+				$kiwi -> failed ();
+				$kiwi -> error  ($data);
+				return;
+			}
 		}
 	}
 	#==========================================
@@ -3488,7 +3519,7 @@ sub installLogicalExtend {
 	#==========================================
 	# dump image file from device if requested
 	#------------------------------------------
-	if ($device) {
+	if (($device) && (! $this->{gdata}->{StudioNode})) {
 		$this -> cleanMount();
 		$name = $this -> buildImageName ();
 		my $dest = $this->{imageDest}."/".$name;
